% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fitPhyloNetwork.R
\name{fitPhyloNetwork}
\alias{fitPhyloNetwork}
\title{Fit a model of network evolution to network data and a phylogenetic tree}
\usage{
fitPhyloNetwork(phy, network, pars)
}
\arguments{
\item{phy}{An object of class 'phylo'}

\item{network}{A network, specified by an n x n matrix filled with 0s and 1s}

\item{pars}{Model parameters: q01, q10, pSpec}
}
\description{
This function returns the likelihood
}

