% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simPhyloNetwork.R
\name{simPhyloNetwork}
\alias{simPhyloNetwork}
\title{Simulate a model of network evolution on a phylogenetic tree}
\usage{
simPhyloNetwork(phy, qRate, sProb)
}
\arguments{
\item{phy}{An object of class 'phylo'}

\item{qRate}{Transition rate q01 = q10}

\item{sProb}{Probability that sister species will interact just after speciation}
}
\description{
This function returns a network with row and column names that match the tree
}

